Huge Portion of Challenge Commit
I apologize but this is going to be a commit of close to the full challenge because I mindlessly didn't fork the repo originally and just created my own repo and did my work there. That repository will be in my github called aircall_calvin so the commits can still be viewed there.